experiments/network-similarity/requirements.txt
torchbearer
torchcam
matplotlib
einops
pandas
